% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{The following functions read, summarize and plot data from the US National Highway
Traffic Safety Administration's Fatality Analysis Reporting System (FARS)
\href{http://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{Fatality Analysis Reporting System (FARS)}.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{the name of the file containing the data to  be read.}
}
\value{
This function returns a data frame table when the file exists. An
error is given if the file doesn't exist. The file has to be located in
the present working directory
}
\description{
Data reading by file name
}
\details{
This function reads a FARS file and creates a data
frame table. The file name is generated by the function make_filename.
}
\examples{
\dontrun{
  my_data <- fars_read("accident_2013.csv.bz2")
}

}
