% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reading multiple years of FARS data files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of integers or strings with the year of the FARS data files}
}
\value{
a list of FARS months and years based on the FARS files specified by
the years input
}
\description{
This function extracts months and years of every FARS
file within the years specified. An error is retuned for invalid years
specified when the FARS file doesn't exits for that year
}
\examples{
  \dontrun{
    my_data <- fars_read_years(c(2013, 2014, 2015))
    my_data <- fars_read_years(c("2013", "2014", "2015"))
  }

}
