% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{File name making}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or string indicating the year of the FARS report}
}
\value{
This function return a string with the FARS file name
}
\description{
This function creates the FARS file name based in the year covered
by the data
}
\examples{
\dontrun{
my_file <- make_filename(2013)
my_file <- make_filename("2013")
}

}
